+++ frontend/src/pages/mod.rs
@@ -1,6 +1,6 @@
 pub mod bookings;
 pub mod make_booking;
+pub mod navbar;
 pub mod observe;
 pub mod telescope;
 pub mod weather;
-pub mod welcome;
